Added the Jupyter Notebook file
Added the Jupyter Notebook file containing the full Python code and Markdown text to help show the project and steps in full. Please ensure to read the README.md file first for the project outline and if you are having difficulty in viewing this Jupyter Notebook file, please view the HTML version instead - I uploaded the HTML version for that reason.